+++ packages/help-center/package.json
@@ -52,11 +52,11 @@
 		"@automattic/calypso-color-schemes": "workspace:^",
 		"@automattic/calypso-typescript-config": "workspace:^",
 		"@storybook/addon-backgrounds": "^7.6.19",
+		"@types/smooch": "^5.3.7",
 		"typescript": "^5.3.3"
 	},
 	"peerDependencies": {
 		"@automattic/calypso-router": "workspace:^",
-		"@types/smooch": "^5.3.7",
 		"@wordpress/data": "^10.8.0",
 		"@wordpress/element": "^6.8.0",
 		"react": "^18.3.1",
