+++ packages/help-center/src/components/help-center-third-party-cookies-notice.scss
@@ -11,14 +11,3 @@
 		padding: 0 30px 20px;
 	}
 }
-
-.help-center__cookie-warning {
-	background-color: var(--studio-yellow-0);
-	border-left: 4px solid var(--studio-yellow-20);
-	padding: 8px 16px;
-	margin-bottom: 30px;
-
-	p {
-		margin: 0;
-	}
-}
