+++ packages/odie-client/package.json
@@ -30,6 +30,7 @@
 	},
 	"dependencies": {
 		"@automattic/components": "workspace:^",
+		"@automattic/i18n-utils": "workspace:^",
 		"@automattic/zendesk-client": "workspace:^",
 		"@tanstack/react-query": "^5.15.5",
 		"@wordpress/api-fetch": "7.8.0",
