+++ packages/odie-client/src/components/message/get-support.scss
@@ -30,6 +30,17 @@
 	}
 }
 
+.help-center__cookie-warning {
+	background-color: var(--studio-yellow-0);
+	border-left: 4px solid var(--studio-yellow-20);
+	padding: 8px 16px;
+	margin-bottom: 30px;
+
+	p {
+		margin: 0;
+	}
+}
+
 .help-center__container-chat .chatbox-messages {
 	.odie__transfer-chat {
 		margin-left: 46px;
