+++ packages/odie-client/src/components/message/get-support.tsx
@@ -1,8 +1,9 @@
-import { NewThirdPartyCookiesNotice } from '@automattic/help-center';
+import { localizeUrl } from '@automattic/i18n-utils';
 import { __ } from '@wordpress/i18n';
 import { useNavigate } from 'react-router-dom';
 import { useOdieAssistantContext } from '../../context';
 import { useCreateZendeskConversation } from '../../hooks';
+
 import './get-support.scss';
 
 interface GetSupportProps {
@@ -14,6 +15,29 @@ interface ButtonConfig {
 	text: string;
 	action: () => Promise< void >;
 }
+
+export const NewThirdPartyCookiesNotice: React.FC = () => {
+	return (
+		<div className="help-center__cookie-warning">
+			<p>
+				<strong>{ __( 'Enable cookies to get support.', __i18n_text_domain__ ) }</strong>
+				&nbsp;
+				{ __(
+					'To access support, please turn on third-party cookies for WordPress.com.',
+					__i18n_text_domain__
+				) }
+				&nbsp;
+				<a
+					target="_blank"
+					rel="noopener noreferrer"
+					href={ localizeUrl( 'https://wordpress.com/support/third-party-cookies/' ) }
+				>
+					{ __( 'Learn more.', __i18n_text_domain__ ) }
+				</a>
+			</p>
+		</div>
+	);
+};
 
 export const GetSupport: React.FC< GetSupportProps > = ( {
 	onClickAdditionalEvent,
